requirements.txt
azure-devops==6.0.0b4
msrest==0.6.21
slack_sdk==3.20.0
